requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.11
djangorestframework==3.13.1
importlib-metadata==4.8.3
kombu==5.1.0
prompt-toolkit==3.0.24
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
